requirements.txt
xxhash
python-docx
simplemma
contractions
flask
easyocr